Give Logistics Teleport Pipe a texture & tooltip
